requirements.txt
asgiref==3.8.1
Django==4.2.18
django-modeltranslation==0.19.12
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
